[PM-30] Project refactor
Class extraction to interfaces and separate responsibilities
The price calculation is done now outside of the cart - and the cart holds information and allows modification against an interface
