+++ .vscode/settings.json
@@ -6,5 +6,6 @@
   "mocha.files.glob": "src/**/*.spec.ts",
   "mocha.options": { "compilers": { "ts": "ts-node/register" } },
   "mocha.requires": ["ts-node/register"],
-  "mocha.runTestsOnSave": "true"
+  "mocha.runTestsOnSave": "true",
+  "eslint.format.enable": true
 }
