+++ src/test/model/promotions/amount-promotion.spec.ts
@@ -1,6 +1,6 @@
 import 'mocha';
 import 'should';
-import {Cart} from '../../../model/cart';
+import {BasicCart} from '../../../model/cart/basic-cart';
 import {Product} from '../../../model/product';
 import {AmountPromotion} from '../../../model/promotions/amount-promotion';
 
@@ -11,7 +11,7 @@ describe('AmountPromotion', () => {
   describe('AmountPromotion(A,3,130)', () => {
     it('should get a discount of 20', () => {
       const promotion = new AmountPromotion(pA, 3, 130);
-      const cart = new Cart();
+      const cart = new BasicCart();
       cart.add(pA, 3);
       promotion.calculateDiscount(cart)[0].should.be.equal(20);
     });
@@ -21,13 +21,13 @@ describe('AmountPromotion', () => {
     const promotion = new AmountPromotion(pB, 2, 45);
 
     it('should get a discount of 15', () => {
-      const cart = new Cart();
+      const cart = new BasicCart();
       cart.add(pB, 2);
       promotion.calculateDiscount(cart)[0].should.be.equal(15);
     });
 
     it('should get a discount of 30', () => {
-      const cart = new Cart();
+      const cart = new BasicCart();
       cart.add(pB, 5);
       promotion.calculateDiscount(cart)[0].should.be.equal(30);
     });
